requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
numpy==1.16.2
opencv-python==4.0.0.21
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
whitenoise==4.1.2
psycopg2==2.7.7
